IOS-3497 Update folder system
